git: Add .gitignore
This libs repo, likewise for the others, did not contain a .gitignore file.

Evidently, it became problematic to commit changes in the main repository due to the artifacts left
over from the build. Adding the .gitignore will help resolve any apparent issues with the
build system as well as eliminate the slight chance of build artificats finding their
inclusion within a commit.

Signed-off-by: Aun-Ali Zaidi <admin@kodeit.net>